+++ ShopsRUs.Core/Orders/Commands/CreateOrderCommand.cs
@@ -9,11 +9,11 @@ namespace ShopsRUs.Core.Orders.Commands
     public class CreateOrderCommand : IRequest<OrderResponse>
     {
         public string ItemName { get; }
-        public decimal Amount { get; }
+        public float Amount { get; }
         public string OrderType { get; }
         public long CustomerId { get; }
 
-        public CreateOrderCommand(decimal amount, string itemName, string orderType, long customerId)
+        public CreateOrderCommand(float amount, string itemName, string orderType, long customerId)
         {
             Amount = amount;
             ItemName = itemName;
