+++ ShopsRUs.Core/Orders/Responses/OrderResponse.cs
@@ -8,9 +8,9 @@ namespace ShopsRUs.Core.Orders.Responses
     {
         public long OrderId { get; set; }
         public string ItemName { get; set; }
-        public decimal Amount { get; set; }
+        public float Amount { get; set; }
         public string OrderType { get; set; }
         public long CustomerId { get; set; }
-        public decimal DiscountedAmount { get; set; }
+        public float DiscountedAmount { get; set; }
     }
 }
